+++ build-gen.sh
@@ -23,7 +23,6 @@ usage()
   echo "-h       displays usage information and exits"
   echo "-i       build interface bindings for SWIG"
   echo "-l       add locale files"
-  echo "-o       build ODBC"
   echo "-p       prepare only, do not run (ninja) after generating build files"
   echo "-s       build samples binaries as well"
   echo "-t       build tests binaries as well"
@@ -40,14 +39,13 @@ option_dir=build
 option_generator="-G Ninja"
 option_github=
 option_locale=
-option_odbc=
 option_prepare=false
 option_samples=
 option_swig=
 option_tests=
 option_tidy=
 
-while getopts ":B:d:D:G:abcghilopstT" opt; do
+while getopts ":B:d:D:G:abcghilpstT" opt; do
   case $opt in
     a)
       option_asan="-DwexENABLE_ASAN=ON"
@@ -94,10 +92,6 @@ while getopts ":B:d:D:G:abcghilopstT" opt; do
 
     l)
       option_locale="-DwexENABLE_GETTEXT=ON"
-    ;;
-
-    o)
-      option_odbc="-DwexENABLE_ODBC=ON"
     ;;
 
     p)
@@ -150,7 +144,6 @@ cmake -B "${option_dir}" "${option_generator}" \
   ${option_coverage} \
   ${option_github} \
   ${option_locale} \
-  ${option_odbc} \
   ${option_samples} \
   ${option_swig} \
   ${option_tests} \