+++ cmake/options.cmake
@@ -1,7 +1,6 @@
 # wex options
 
 option(wexENABLE_GETTEXT "Use gettext for i18n")
-option(wexENABLE_ODBC "Use ODBC class")
 
 option(wexBUILD_SAMPLES "Build samples")
 option(wexBUILD_SHARED "Build shared libs")