+++ cmake/Findwex.cmake
@@ -52,14 +52,6 @@ find_package(
   REQUIRED
 )
 
-find_package(ODBC QUIET)
-
-if(ODBC_FOUND)
-  add_definitions(-DwexUSE_ODBC)
-else()
-  set(ODBC_LIBRARIES "")
-endif()
-
 if(WIN32)
   add_definitions(-D__WXMSW__)
 
@@ -244,7 +236,6 @@ if(${wex_FOUND})
     ${apple_LIBRARIES}
     ${Boost_LIBRARIES}
     ${cpp_LIBRARIES}
-    ${ODBC_LIBRARIES}
   )
 else()
   message(FATAL_ERROR "No suitable wex found in: " ${wex_INCLUDES})