+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - editor pages can be pinned to prevent moving them
 - added more vim g and z commands
+- added showing blame current line in statusbar
+- added pane_is_shown to statusbar
 
 ### Changed
 
@@ -21,6 +23,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - build-gen.sh supports xcodebuild
 - updated git grep dialog to include checkboxes
 - more boost::process::v2 lib used, requiring boost 1.88
+- used released wxWidgets 3.3.0
+- allow to enable a log_none
 
 ### Fixed
 
@@ -37,6 +41,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Removed
 
+- removed ODBC support and otlv4 lib: moved to wex-db lib
 - removed queue_thread
 
 ## [25.04] - 2025-03-16